+++ R/zzz.R
@@ -14,6 +14,10 @@
   )
   toset <- !(names(op.fmridataset) %in% names(op))
   if (any(toset)) options(op.fmridataset[toset])
+  
+  if (.require_namespace("DelayedArray", quietly = TRUE)) {
+    register_delayed_array_support()
+  }
 
   invisible()
 }